developpement/project/requirements_api.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.1.2+cpu ; sys_platform != 'darwin'
torch==2.1.2 ; sys_platform == 'darwin'

torchvision==0.16.2+cpu ; sys_platform != 'darwin'
torchvision==0.16.2 ; sys_platform == 'darwin'

Flask==2.3.2
Pillow==10.0.0
numpy==1.24.4
matplotlib==3.7.2
annoy==1.17.3
transformers==4.21.1
datasets==2.0.0
scikit-learn==1.3.2
nltk==3.8.1